+++ WebMVC/Controllers/StoreController.cs
@@ -15,27 +15,37 @@ namespace WebMVC.Controllers
 
         public ActionResult Create()
         {
-            return View();
+            if (IsAdmin())
+            {
+                return View();
+            }
+
+            return RedirectToAction("Index");
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(StoreCreate model)
         {
-            if (!ModelState.IsValid)
+            if (IsAdmin())
             {
+                if (!ModelState.IsValid)
+                {
+                    return View(model);
+                }
+
+                var service = CreateService();
+
+                if (service.CreateStore(model))
+                {
+                    TempData["SaveResult"] = "Your Store was created.";
+                    return RedirectToAction("Index");
+                };
+
                 return View(model);
             }
 
-            var service = CreateService();
-
-            if (service.CreateStore(model))
-            {
-                TempData["SaveResult"] = "Your Store was created.";
-                return RedirectToAction("Index");
-            };
-
-            return View(model);
+            return RedirectToAction("Index");
         }
 
         public ActionResult Details(int id)
@@ -48,57 +58,72 @@ namespace WebMVC.Controllers
 
         public ActionResult Edit(int id)
         {
-            var service = CreateService();
-            var detail = service.GetStoreById(id);
-            var model =
-                new StoreEdit
-                {
-                    StoreId = detail.StoreId,
-                    StoreName = detail.StoreName,
-                    StoreStreet = detail.StoreStreet,
-                    StoreCity = detail.StoreCity,
-                    StoreState = detail.StoreState,
-                    StoreZip = detail.StoreZip,
-                    StorePhoneNumber = detail.StorePhoneNumber,
-                    OwnerId = detail.OwnerId
-                };
-            return View(model);
+            if (IsAdmin())
+            {
+                var service = CreateService();
+                var detail = service.GetStoreById(id);
+                var model =
+                    new StoreEdit
+                    {
+                        StoreId = detail.StoreId,
+                        StoreName = detail.StoreName,
+                        StoreStreet = detail.StoreStreet,
+                        StoreCity = detail.StoreCity,
+                        StoreState = detail.StoreState,
+                        StoreZip = detail.StoreZip,
+                        StorePhoneNumber = detail.StorePhoneNumber,
+                        OwnerId = detail.OwnerId
+                    };
+                return View(model);
+            }
+
+            return RedirectToAction("Index");
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(int id, StoreEdit model)
         {
-            if (!ModelState.IsValid)
+            if (IsAdmin())
             {
+                if (!ModelState.IsValid)
+                {
+                    return View(model);
+                }
+
+                if (model.StoreId != id)
+                {
+                    ModelState.AddModelError("", "Id Mismatch");
+                    return View(model);
+                }
+
+                var service = CreateService();
+
+                if (service.UpdateStore(model))
+                {
+                    TempData["SaveResult"] = "Your Store was updated.";
+                    return RedirectToAction("Index");
+                }
+
+                ModelState.AddModelError("", "Your Store could not be updated.");
                 return View(model);
             }
 
-            if (model.StoreId != id)
-            {
-                ModelState.AddModelError("", "Id Mismatch");
-                return View(model);
-            }
-
-            var service = CreateService();
-
-            if (service.UpdateStore(model))
-            {
-                TempData["SaveResult"] = "Your Store was updated.";
-                return RedirectToAction("Index");
-            }
-
-            ModelState.AddModelError("", "Your Store could not be updated.");
-            return View(model);
+            return RedirectToAction("Index");
         }
 
         [ActionName("Delete")]
         public ActionResult Delete(int id)
         {
-            var svc = CreateService();
-            var model = svc.GetStoreById(id);
+            if (IsAdmin())
+            {
+                var svc = CreateService();
+                var model = svc.GetStoreById(id);
 
-            return View(model);
+                return View(model);
+            }
+
+            return RedirectToAction("Index");
         }
 
         [HttpPost]
@@ -106,11 +131,16 @@ namespace WebMVC.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeletePost(int id)
         {
-            var service = CreateService();
+            if (IsAdmin())
+            {
+                var service = CreateService();
 
-            service.DeleteStore(id);
+                service.DeleteStore(id);
 
-            TempData["SaveResult"] = "Your Store was deleted";
+                TempData["SaveResult"] = "Your Store was deleted";
+
+                return RedirectToAction("Index");
+            }
 
             return RedirectToAction("Index");
         }
@@ -118,6 +148,10 @@ namespace WebMVC.Controllers
         private StoreService CreateService()
         {
             return new StoreService(Guid.Parse(User.Identity.GetUserId()));
+        }
+        private bool IsAdmin()
+        {
+            return User.IsInRole("Admin");
         }
     }
 }
